Added LogLineReadCloser and split up files
