+++ pkg/resultstore/logline_test.go
@@ -1,7 +1,6 @@
 package resultstore
 
 import (
-	"bytes"
 	"io"
 	"testing"
 	"time"
@@ -9,34 +8,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestLogLineWriteCloser(t *testing.T) {
-	var buf bytes.Buffer
-	var w LogLineWriteCloser = logLineWriteCloser{
-		writeCloser: nopWriteCloser{&buf},
-		store:       &store{},
-	}
-	err := w.WriteLogLine(sampleTimeStr + " Foo bar")
-	require.NoError(t, err)
-
-	want := sampleTimeStr + " Foo bar\n"
-	got := buf.String()
-	assert.Equal(t, want, got)
-}
-
-func TestLogLineWriteCloser_Sanitizes(t *testing.T) {
-	var buf bytes.Buffer
-	var w LogLineWriteCloser = logLineWriteCloser{
-		writeCloser: nopWriteCloser{&buf},
-		store:       &store{},
-	}
-	err := w.WriteLogLine(sampleTimeStr + " Foo \nbar")
-	require.NoError(t, err)
-
-	want := sampleTimeStr + " Foo \\nbar\n"
-	got := buf.String()
-	assert.Equal(t, want, got)
-}
 
 func TestParseLogLine(t *testing.T) {
 	zeroTime := time.Time{}
